requirements.txt
anyio==4.0.0
certifi==2023.11.17
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
psycopg2==2.9.9
python-dotenv==1.0.0
python-telegram-bot==20.6
sniffio==1.3.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
